requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
faiss-cpu
sentence-transformers 
streamlit
requests